chore(core): reorganize config files
